+++ vllm/model_executor/layers/quantization/compressed_tensors/schemes/__init__.py
@@ -2,18 +2,35 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
 from .compressed_tensors_scheme import CompressedTensorsScheme
-from .compressed_tensors_w4a4_nvfp4 import CompressedTensorsW4A4Fp4
+from .compressed_tensors_w4a4_nvfp4 import (
+    CompressedTensorsW4A4Fp4,
+    CompressedTensorsW4A4Nvfp4MoeMethod,
+)
 from .compressed_tensors_w4a8_fp8 import CompressedTensorsW4A8Fp8
-from .compressed_tensors_w4a8_int import CompressedTensorsW4A8Int
+from .compressed_tensors_w4a8_int import (
+    CompressedTensorsW4A8Int,
+    CompressedTensorsW4A8Int8MoEMethod,
+)
 from .compressed_tensors_w4a16_24 import (
     W4A16SPARSE24_SUPPORTED_BITS,
     CompressedTensorsW4A16Sparse24,
 )
 from .compressed_tensors_w4a16_nvfp4 import CompressedTensorsW4A16Fp4
-from .compressed_tensors_w8a8_fp8 import CompressedTensorsW8A8Fp8
-from .compressed_tensors_w8a8_int8 import CompressedTensorsW8A8Int8
+from .compressed_tensors_w8a8_fp8 import (
+    CompressedTensorsW8A8Fp8,
+    CompressedTensorsW8A8Fp8MoEMethod,
+)
+from .compressed_tensors_w8a8_int8 import (
+    CompressedTensorsW8A8Int8,
+    CompressedTensorsW8A8Int8MoEMethod,
+)
 from .compressed_tensors_w8a16_fp8 import CompressedTensorsW8A16Fp8
-from .compressed_tensors_wNa16 import WNA16_SUPPORTED_BITS, CompressedTensorsWNA16
+from .compressed_tensors_wNa16 import (
+    WNA16_SUPPORTED_BITS,
+    CompressedTensorsWNA16,
+    CompressedTensorsWNA16MarlinMoEMethod,
+    CompressedTensorsWNA16MoEMethod,
+)
 
 # This avoids circular import error
 from .compressed_tensors_24 import CompressedTensors24  # isort: skip
@@ -32,4 +49,11 @@ __all__ = [
     "CompressedTensorsW4A4Fp4",
     "CompressedTensorsW4A8Int",
     "CompressedTensorsW4A8Fp8",
+    # MoE methods
+    "CompressedTensorsW4A4Nvfp4MoeMethod",
+    "CompressedTensorsW8A8Fp8MoEMethod",
+    "CompressedTensorsW8A8Int8MoEMethod",
+    "CompressedTensorsWNA16MarlinMoEMethod",
+    "CompressedTensorsWNA16MoEMethod",
+    "CompressedTensorsW4A8Int8MoEMethod",
 ]
